+++ client/src/views/Directory/Malt.vue
@@ -2,21 +2,61 @@
   <div>
     <div class="row">
       <div class="col s12">
-        <h2>Солод <small>Справочник</small></h2>
+        <h4>Солод <small class="grey-text">Справочник</small></h4>
       </div>
     </div>
-    <table-find :headers="headers" :data="MALT_LIST" @add-btn-click="addBtnClick" @table-btn-click="tableBtnClick"></table-find>
-    <modal title="Солод" :component="maltForm" :check-validate="true" @result-modal="resultModal" ref="modal"></modal>
+    <table-component
+      :headers="headers"
+      :counter="true"
+      :min-length-find="1"
+      :finder="true"
+      :rows-data="MALT_LIST"
+      :rows-on-page-list="[10, 25, 50, 75, 100]"
+      :start-rows-on-page="50"
+      :pages="true"
+      :checkbox="false"
+      :row-buttons="[
+        {
+          icon: 'edit',
+          class: 'btn-small waves-effect white black-text z-depth-0',
+          result: 'edit',
+        },
+        {
+          icon: 'trash',
+          class: 'btn-small waves-effect white black-text z-depth-0',
+          result: 'delete',
+        },
+      ]"
+      :table-buttons="[
+        {
+          title: 'Добавить',
+          icon: 'plus',
+          iconClass: 'left',
+          result: 'add',
+          class: 'btn-small waves-effect',
+        },
+      ]"
+      @row-btn-click="rowBtnClick"
+      @table-btn-click="tableBtnClick"
+    />
+    <modal
+      title="Солод"
+      :component="maltForm"
+      :check-validate="true"
+      @result-modal="resultModal"
+      ref="modal"
+    ></modal>
   </div>
 </template>
 
 <script>
 import { mapGetters, mapActions, mapMutations } from 'vuex'
 import Modal from '../../components/Modal.vue'
-import TableFind from '../../components/TableFind.vue'
 import MaltForm from '../../components/DirectoryForm/MaltForm.vue'
+import TableComponent from '../../components/TableComponent.vue'
+
 export default {
-  components: { Modal, MaltForm, TableFind },
+  components: { Modal, MaltForm, TableComponent },
   computed: {
     ...mapGetters(['MALT_LIST', 'MALT']),
   },
@@ -33,9 +73,11 @@ export default {
   methods: {
     ...mapActions(['getMaltList', 'getMaltById', 'postMalt', 'putMaltById', 'deleteMaltById']),
     ...mapMutations(['CLR_MALT']),
-    addBtnClick() {
-      this.CLR_MALT()
-      this.$refs.modal.open()
+    tableBtnClick(e) {
+      if (e.button == 'add') {
+        this.CLR_MALT()
+        this.$refs.modal.open()
+      }
     },
     resultModal(e) {
       if (!e.result) {
@@ -55,8 +97,8 @@ export default {
 
       this.$refs.modal.close()
     },
-    tableBtnClick(e) {
-      if (e.button === 'delete') {
+    rowBtnClick(e) {
+      if (e.button === 'delete' && confirm('Вы действительно хотите удалить элемент?')) {
         this.deleteMaltById(e.data.id)
       }
       if (e.button === 'edit') {
