+++ client/src/views/Directory/OtherIngredients.vue
@@ -2,11 +2,50 @@
   <div>
     <div class="row">
       <div class="col s12">
-        <h2>Другие ингредиенты <small>Справочник</small></h2>
+        <h4>Другие ингредиенты <small class="grey-text">Справочник</small></h4>
       </div>
     </div>
-    <table-find :headers="headers" :data="OTHER_INGREDIENT_LIST" @add-btn-click="addBtnClick" @table-btn-click="tableBtnClick"></table-find>
-    <modal title="Другие ингредиенты" :component="otherIngredientsForm" :check-validate="true" @result-modal="resultModal" ref="modal"></modal>
+    <table-component
+      :headers="headers"
+      :counter="true"
+      :min-length-find="1"
+      :finder="true"
+      :rows-data="OTHER_INGREDIENT_LIST"
+      :rows-on-page-list="[10, 25, 50, 75, 100]"
+      :start-rows-on-page="50"
+      :pages="true"
+      :checkbox="false"
+      :row-buttons="[
+        {
+          icon: 'edit',
+          class: 'btn-small waves-effect white black-text z-depth-0',
+          result: 'edit',
+        },
+        {
+          icon: 'trash',
+          class: 'btn-small waves-effect white black-text z-depth-0',
+          result: 'delete',
+        },
+      ]"
+      :table-buttons="[
+        {
+          title: 'Добавить',
+          icon: 'plus',
+          iconClass: 'left',
+          result: 'add',
+          class: 'btn-small waves-effect',
+        },
+      ]"
+      @row-btn-click="rowBtnClick"
+      @table-btn-click="tableBtnClick"
+    />
+    <modal
+      title="Другие ингредиенты"
+      :component="otherIngredientsForm"
+      :check-validate="true"
+      @result-modal="resultModal"
+      ref="modal"
+    ></modal>
   </div>
 </template>
 
@@ -14,10 +53,10 @@
 import { mapGetters, mapActions, mapMutations } from 'vuex'
 import Modal from '../../components/Modal.vue'
 import OtherIngredientsForm from '../../components/DirectoryForm/OtherIngredientsForm.vue'
-import TableFind from '../../components/TableFind.vue'
+import TableComponent from '../../components/TableComponent.vue'
 
 export default {
-  components: { OtherIngredientsForm, Modal, TableFind },
+  components: { OtherIngredientsForm, Modal, TableComponent },
   computed: {
     ...mapGetters(['OTHER_INGREDIENT_LIST', 'OTHER_INGREDIENT']),
   },
@@ -32,11 +71,19 @@ export default {
     ],
   }),
   methods: {
-    ...mapActions(['getOtherIngredientList', 'getOtherIngredientById', 'postOtherIngredient', 'putOtherIngredientById', 'deleteOtherIngredientById']),
+    ...mapActions([
+      'getOtherIngredientList',
+      'getOtherIngredientById',
+      'postOtherIngredient',
+      'putOtherIngredientById',
+      'deleteOtherIngredientById',
+    ]),
     ...mapMutations(['CLR_OTHER_INGREDIENT']),
-    addBtnClick() {
-      this.CLR_OTHER_INGREDIENT()
-      this.$refs.modal.open()
+    tableBtnClick(e) {
+      if (e.button == 'add') {
+        this.CLR_OTHER_INGREDIENT()
+        this.$refs.modal.open()
+      }
     },
     resultModal(e) {
       if (!e.result) {
@@ -56,8 +103,8 @@ export default {
 
       this.$refs.modal.close()
     },
-    tableBtnClick(e) {
-      if (e.button === 'delete') {
+    rowBtnClick(e) {
+      if (e.button === 'delete' && confirm('Вы действительно хотите удалить элемент?')) {
         this.deleteOtherIngredientById(e.data.id)
       }
       if (e.button === 'edit') {
